+++ resources/js/react/index.ts
@@ -1,0 +1,3 @@
+export { Ga4Snippet } from './Ga4Snippet'
+export { GaOverview } from './GaOverview'
+export type { GaOverviewProps, GaOverviewData } from './GaOverview'